refactor: simplify Ted Intorcio bio component
