+++ src/components/common/NavItems.ts
@@ -1,0 +1,9 @@
+export const NAV_ITEMS = [
+  { label: "HOME", path: "/" },
+  { label: "ABOUT US", path: "/about-us" },
+  { label: "OUR WORK", path: "/our-work" },
+  { label: "OUR SERVICES", path: "/services" },
+  { label: "CONTACT US", path: "/contact-us" },
+] as const;
+
+export const QUOTE_ITEM = { label: "GET QUOTE", path: "/contact-us" } as const;
