+++ src/components/pages/services/OurServices.tsx
@@ -3,6 +3,7 @@ import { Outlet } from "react-router-dom"
 import { OurServicesCard1 } from "./OurServicesCard1"
 import { OurServicesCard2 } from "./OurServicesCard2"
 import ourServicesBanner from "/src/assets/services/services-banner.png"
+import { HeroBanner } from "../../common/HeroBanner"
 import construction from "/src/assets/services/construction.png"
 import tourism from "/src/assets/services/tourism.jpg"
 import corporate from "/src/assets/services/corporate.jpeg"
@@ -22,11 +23,7 @@ export const OurServices: React.FC<Props> = () => (
     </Helmet>
 
     <div className="flex flex-col items-center justify-center">
-      <img
-        src={ourServicesBanner}
-        className="h-32 w-full object-cover object-top sm:h-96"
-        fetchPriority="high"
-      />
+      <HeroBanner img={ourServicesBanner} />
 
       <div className="flex max-w-[70em] flex-col gap-4 px-6 py-20">
         <p className="text-xl text-blue-500">OUR&nbsp;SERVICES</p>
